+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* @dmlap Check a component is a function before new-ing ([view](https://github.com/videojs/video.js/pull/2814))
 
 --------------------